+++ frontend/src/pages/ComplianceRequestPage.tsx
@@ -1,23 +1,12 @@
 import {
-  Badge,
   Button,
-  Callout,
-  Checkbox,
   Flex,
   Heading,
-  Link,
-  Select,
-  Table,
   Tabs,
-  Text,
-  TextArea,
-  TextField,
-  Tooltip,
 } from '@radix-ui/themes';
-import { IconAlertTriangle } from '@tabler/icons-react';
 import { useMutation, useQuery } from '@tanstack/react-query';
 import { useEffect, useMemo, useRef, useState } from 'react';
-import { Link as RouterLink, useLocation, useNavigate, useSearchParams } from 'react-router-dom';
+import { useLocation, useNavigate, useSearchParams } from 'react-router-dom';
 
 import {
   createComplianceRequest,
@@ -36,25 +25,15 @@ import { ErrorAlert } from '../components/ErrorAlert';
 import { PageSpinner } from '../components/PageSpinner';
 import { useConfirm } from '../components/useConfirm';
 import { useDocumentTitle } from '../hooks/useDocumentTitle';
+import StandardStep from '../components/StandardStep';
+import RunSelectionStep from '../components/RunSelectionStep';
+import { Mode, FormState } from '../utils/complianceRequestWizard';
+import DerDetailsStep from '../components/DerDetailsStep';
+import SoftwareClientDetailsStep from '../components/SoftwareClientDetailsStep';
 
-type Mode = 'new' | 'edit' | 'view';
 
-const STEP_TITLES = ['Standard', 'Classes & Runs', 'DER', 'Software Client'];
+const STEP_TITLES = ['Compliance Details', 'Run Selection', 'DER Details', 'Software Client Details'];
 
-interface FormState {
-  csip_aus_version: string;
-  witnessed_at: string;
-  classes: Set<string>;
-  runByProcedure: Record<string, number>;
-  der_brand: string;
-  der_oem: string;
-  der_series: string;
-  der_representative_models: string;
-  software_client_type: string;
-  software_client_providers: string;
-  software_client_versions: string;
-  onsite_hardware_details: string;
-}
 
 function emptyForm(): FormState {
   return {
@@ -276,7 +255,7 @@ export function ComplianceRequestPage({ isAdminView }: { isAdminView: boolean })
         </Tabs.Content>
 
         <Tabs.Content value="1">
-          <ClassesRunsStep
+          <RunSelectionStep
             form={form}
             readOnly={readOnly}
             formData={formData}
@@ -293,11 +272,11 @@ export function ComplianceRequestPage({ isAdminView }: { isAdminView: boolean })
         </Tabs.Content>
 
         <Tabs.Content value="2">
-          <DerStep form={form} readOnly={readOnly} update={update} />
+          <DerDetailsStep form={form} readOnly={readOnly} update={update} />
         </Tabs.Content>
 
         <Tabs.Content value="3">
-          <SoftwareClientStep form={form} readOnly={readOnly} update={update} />
+          <SoftwareClientDetailsStep form={form} readOnly={readOnly} update={update} />
         </Tabs.Content>
       </Tabs.Root>
 
@@ -415,364 +394,4 @@ function buildInitialForm(
     }
   }
   return form;
-}
-
-function StandardStep({
-  form,
-  mode,
-  readOnly,
-  versions,
-  update,
-}: {
-  form: FormState;
-  mode: Mode;
-  readOnly: boolean;
-  versions: string[];
-  update: (patch: Partial<FormState>) => void;
-}) {
-  return (
-    <Flex direction="column" gap="4" pt="4">
-      {mode === 'new' && (
-        <Callout.Root color="yellow">
-          <Callout.Icon>
-            <IconAlertTriangle size={16} />
-          </Callout.Icon>
-          <Callout.Text>
-            Witness testing must be completed before a compliance request can be submitted. If you
-            are yet to complete witness testing, please contact{' '}
-            <Link href="mailto:support@bsgip.com">support@bsgip.com</Link> to arrange a date.
-          </Callout.Text>
-        </Callout.Root>
-      )}
-      <label>
-        <Text as="div" size="2" weight="bold" mb="1">
-          Compliance Standard
-        </Text>
-        <Select.Root
-          value={form.csip_aus_version}
-          onValueChange={(v) => update({ csip_aus_version: v })}
-          disabled={readOnly}
-        >
-          <Select.Trigger placeholder="Select a version" />
-          <Select.Content>
-            {versions.map((v) => (
-              <Select.Item key={v} value={v}>
-                CSIP-Aus {v}
-              </Select.Item>
-            ))}
-          </Select.Content>
-        </Select.Root>
-        <Text as="div" size="1" color="gray" mt="1">
-          A compliance request can only be made against one compliance standard at a time.
-        </Text>
-      </label>
-      <label>
-        <Text as="div" size="2" weight="bold" mb="1">
-          Witness Testing Date
-        </Text>
-        <TextField.Root
-          type="date"
-          value={form.witnessed_at}
-          onChange={(e) => update({ witnessed_at: e.target.value })}
-          disabled={readOnly}
-        />
-        <Text as="div" size="1" color="gray" mt="1">
-          The date when in-person witness testing was performed.
-        </Text>
-      </label>
-    </Flex>
-  );
-}
-
-function ClassesRunsStep({
-  form,
-  readOnly,
-  formData,
-  classesForVersion,
-  visibleProcedures,
-  runsByProcedure,
-  missingByClass,
-  missingCount,
-  toggleClass,
-  setAllClasses,
-  update,
-  isAdminView,
-}: {
-  form: FormState;
-  readOnly: boolean;
-  formData: ComplianceFormDataResponse;
-  classesForVersion: string[];
-  visibleProcedures: string[];
-  runsByProcedure: Record<string, RunResponse[]>;
-  missingByClass: Record<string, string[]>;
-  missingCount: number;
-  toggleClass: (c: string, checked: boolean) => void;
-  setAllClasses: (checked: boolean) => void;
-  update: (patch: Partial<FormState>) => void;
-  isAdminView: boolean;
-}) {
-  const descriptions = useMemo(
-    () => new Map(formData.compliance_classes.map((c) => [c.name, c.description])),
-    [formData.compliance_classes]
-  );
-
-  return (
-    <Flex direction="column" gap="4" pt="4">
-      <Flex direction="column" gap="2">
-        <Heading as="h3" size="4">
-          Classes
-        </Heading>
-        <Text size="2">Choose all the compliance classes you want to be assessed under.</Text>
-        {!readOnly && (
-          <Flex gap="2">
-            <Button size="1" onClick={() => setAllClasses(true)}>
-              Select All
-            </Button>
-            <Button size="1" variant="soft" color="gray" onClick={() => setAllClasses(false)}>
-              Deselect All
-            </Button>
-          </Flex>
-        )}
-        <Flex gap="3" wrap="wrap" mt="1">
-          {classesForVersion.map((c) => (
-            <Tooltip key={c} content={descriptions.get(c) || c}>
-              <Text as="label" size="2" style={{ minWidth: 180 }}>
-                <Flex gap="2" align="center">
-                  <Checkbox
-                    checked={form.classes.has(c)}
-                    onCheckedChange={(checked) => toggleClass(c, checked === true)}
-                    disabled={readOnly}
-                  />
-                  {c}
-                </Flex>
-              </Text>
-            </Tooltip>
-          ))}
-        </Flex>
-      </Flex>
-
-      {missingCount > 0 && (
-        <Callout.Root color="red">
-          <Callout.Icon>
-            <IconAlertTriangle size={16} />
-          </Callout.Icon>
-          <Callout.Text>
-            <Text as="div" weight="bold" mb="1">
-              Runs Missing
-            </Text>
-            Some chosen compliance classes do not have successful test runs:
-            <ul style={{ margin: '8px 0' }}>
-              {Object.entries(missingByClass).map(([c, missing]) => (
-                <li key={c}>
-                  <strong>Class {c}:</strong> {missing.join(', ')}
-                </li>
-              ))}
-            </ul>
-            There is a total of <strong>{missingCount}</strong> missing runs. You must complete the
-            required tests or remove the incomplete classes before{' '}
-            {isAdminView ? 'finalising' : 'submitting'}.
-          </Callout.Text>
-        </Callout.Root>
-      )}
-
-      {visibleProcedures.length > 0 && (
-        <Flex direction="column" gap="2">
-          <Heading as="h3" size="4">
-            Runs
-          </Heading>
-          <Text size="2">
-            For each test procedure, choose which run you want assessed. Only successful (finalised
-            and passing) runs are shown.
-          </Text>
-          <Table.Root variant="surface">
-            <Table.Body>
-              {visibleProcedures.map((p) => (
-                <Table.Row key={p}>
-                  <Table.RowHeaderCell>
-                    <Link asChild>
-                      <RouterLink to={`/procedure/${p}`}>{p}</RouterLink>
-                    </Link>
-                  </Table.RowHeaderCell>
-                  <Table.Cell>
-                    <Select.Root
-                      value={form.runByProcedure[p] ? String(form.runByProcedure[p]) : undefined}
-                      onValueChange={(v) =>
-                        update({ runByProcedure: { ...form.runByProcedure, [p]: Number(v) } })
-                      }
-                      disabled={readOnly}
-                    >
-                      <Select.Trigger placeholder="Select a run" />
-                      <Select.Content>
-                        {(runsByProcedure[p] ?? []).map((run) => (
-                          <Select.Item key={run.run_id} value={String(run.run_id)}>
-                            #{run.run_id}
-                          </Select.Item>
-                        ))}
-                      </Select.Content>
-                    </Select.Root>
-                  </Table.Cell>
-                  <Table.Cell>
-                    {form.runByProcedure[p] && (
-                      <Link asChild>
-                        <RouterLink to={`/run/${form.runByProcedure[p]}`}>View</RouterLink>
-                      </Link>
-                    )}
-                  </Table.Cell>
-                </Table.Row>
-              ))}
-            </Table.Body>
-          </Table.Root>
-        </Flex>
-      )}
-      {activeClassesEmpty(form, classesForVersion) && (
-        <Badge color="gray">Select at least one compliance class to choose runs.</Badge>
-      )}
-    </Flex>
-  );
-}
-
-function activeClassesEmpty(form: FormState, classesForVersion: string[]): boolean {
-  return !classesForVersion.some((c) => form.classes.has(c));
-}
-
-function FieldRow({
-  label,
-  help,
-  children,
-}: {
-  label: string;
-  help?: string;
-  children: React.ReactNode;
-}) {
-  return (
-    <label>
-      <Text as="div" size="2" weight="bold" mb="1">
-        {label}
-      </Text>
-      {children}
-      {help && (
-        <Text as="div" size="1" color="gray" mt="1">
-          {help}
-        </Text>
-      )}
-    </label>
-  );
-}
-
-function DerStep({
-  form,
-  readOnly,
-  update,
-}: {
-  form: FormState;
-  readOnly: boolean;
-  update: (patch: Partial<FormState>) => void;
-}) {
-  return (
-    <Flex direction="column" gap="4" pt="4">
-      <Heading as="h3" size="4">
-        DER
-      </Heading>
-      <FieldRow label="Brand">
-        <TextField.Root
-          value={form.der_brand}
-          onChange={(e) => update({ der_brand: e.target.value })}
-          disabled={readOnly}
-        />
-      </FieldRow>
-      <FieldRow label="OEM">
-        <TextField.Root
-          value={form.der_oem}
-          onChange={(e) => update({ der_oem: e.target.value })}
-          disabled={readOnly}
-        />
-      </FieldRow>
-      <FieldRow label="Series" help="The series or product line the DER belongs to.">
-        <TextField.Root
-          value={form.der_series}
-          onChange={(e) => update({ der_series: e.target.value })}
-          disabled={readOnly}
-        />
-      </FieldRow>
-      <FieldRow
-        label="Representative Models"
-        help="Models in the DER series. Models must use the same software and hardware configuration; they may differ in rated power or phases."
-      >
-        <TextArea
-          rows={3}
-          value={form.der_representative_models}
-          onChange={(e) => update({ der_representative_models: e.target.value })}
-          disabled={readOnly}
-        />
-      </FieldRow>
-    </Flex>
-  );
-}
-
-function SoftwareClientStep({
-  form,
-  readOnly,
-  update,
-}: {
-  form: FormState;
-  readOnly: boolean;
-  update: (patch: Partial<FormState>) => void;
-}) {
-  return (
-    <Flex direction="column" gap="4" pt="4">
-      <Heading as="h3" size="4">
-        Software Client
-      </Heading>
-      <FieldRow
-        label="Type"
-        help="Direct - the DER connects with its own client representing a single site. Proxy (Aggregator) - DER(s) connect via a communications aggregator that may represent many sites."
-      >
-        <Select.Root
-          value={form.software_client_type}
-          onValueChange={(v) => update({ software_client_type: v })}
-          disabled={readOnly}
-        >
-          <Select.Trigger />
-          <Select.Content>
-            <Select.Item value="direct">Direct</Select.Item>
-            <Select.Item value="proxy">Proxy (Aggregator)</Select.Item>
-          </Select.Content>
-        </Select.Root>
-      </FieldRow>
-      <FieldRow
-        label="Provider(s)"
-        help="The software client may be provided by an OEM or third-party."
-      >
-        <TextField.Root
-          value={form.software_client_providers}
-          onChange={(e) => update({ software_client_providers: e.target.value })}
-          disabled={readOnly}
-        />
-      </FieldRow>
-      <FieldRow
-        label="Version(s)"
-        help="The full name and version(s) of all software clients, plus any intermediate cloud server(s) or platform(s)."
-      >
-        <TextField.Root
-          value={form.software_client_versions}
-          onChange={(e) => update({ software_client_versions: e.target.value })}
-          disabled={readOnly}
-        />
-      </FieldRow>
-      <Heading as="h3" size="4">
-        On-site Hardware
-      </Heading>
-      <FieldRow
-        label="Details"
-        help="Makes and models of any on-site gateway, control, EMS device or external accessories that are part of the software client implementation."
-      >
-        <TextArea
-          rows={3}
-          value={form.onsite_hardware_details}
-          onChange={(e) => update({ onsite_hardware_details: e.target.value })}
-          disabled={readOnly}
-        />
-      </FieldRow>
-    </Flex>
-  );
 }
